WIP: creates Warrior Instance and Faerie instance, still wokring on aesthetics of app

+++ src/Faerie.java
@@ -7,15 +7,15 @@ import java.util.ArrayList;
  *******************************************/
 
 public class Faerie extends Hero {
-    int height;
+    double height;
 
-    public Faerie(String name, int strength, String health, ArrayList<Weapon> weapons, int height)
+    public Faerie(String name, int strength, String health, ArrayList<Weapon> weapons, double height)
             throws HeroStrengthException {
         super(name, strength, health, weapons);
         this.height = height;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
